apps/pdf-parser/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
tabula-py==2.9.0
camelot-py[cv]==0.11.0
pandas==2.1.4
numpy==1.26.3
opencv-python==4.9.0.80
ghostscript==0.7
pydantic==2.5.3
